requirements.txt
langgraph>=0.2.0
langchain-core>=0.3.0
openai>=1.0.0
python-binance>=1.0.19
pandas>=2.0.0
ta>=0.11.0
python-dotenv>=1.0.0
requests>=2.31.0
yfinance>=0.2.0
flask>=3.0.0
flask-cors>=4.0.0

